+++ apps/api/services/branch_service.py
@@ -3,7 +3,7 @@ from apps.core.models import Branch
 from .company_services import verification_owner_company
 
 
-def get_branch_by_branch_id(user, branch_id):
+def get_branch_by_id(user, branch_id):
     """
     Функция возвращает филиал по его id. Если текущий пользователь не владелец этого филиала, вернется False
     :param user: Текущий пользователь
@@ -18,5 +18,11 @@ def get_branch_by_branch_id(user, branch_id):
 
 
 def verification_owner_branch(user, branch_id):
+    """
+    Функция проверяет, имеет ли текущий пользователь доступ к branch
+    :param user: текущий пользователь
+    :param branch_id: id branch
+    :return: True или False
+    """
     branch = Branch.objects.get(pk=branch_id)
     return verification_owner_company(user=user, company_id=branch.company_id)