+++ apps/api/views/company_views.py
@@ -43,11 +43,7 @@ from ..services.company_services import get_company_by_id, create_company_by_com
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
 def create_company(request):
-    """
-    Контроллер для создания компании
-    :param request:
-    :return: response
-    """
+    """Контроллер для создания компании"""
     user = request.user
 
     try:
@@ -56,7 +52,7 @@ def create_company(request):
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -102,7 +98,7 @@ def read_company(request, pk):
         return Response(data={'detail': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -160,7 +156,7 @@ def update_company(request, pk):
         return Response(data={'detail': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -206,5 +202,5 @@ def delete_company(request, pk):
         return Response(data={'detail': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)