+++ apps/api/views/review_settings_views.py
@@ -56,11 +56,7 @@ from ..services.review_settings_service import get_review_settings_by_id, create
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
 def create_review_settings(request):
-    """
-    Контроллер для создания review settings
-    :param request:
-    :return: response
-    """
+    """Контроллер для создания review settings """
     user = request.user
 
     try:
@@ -80,7 +76,7 @@ def create_review_settings(request):
         return Response(data={'detail': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -125,7 +121,7 @@ def read_review_settings(request, pk):
         return Response(data={'detail': 'Такой review settings не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -184,7 +180,7 @@ def update_review_settings(request, pk):
         return Response(data={'detail': 'Такой review settings не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -229,5 +225,5 @@ def delete_review_settings(request, pk):
         return Response(data={'detail': 'Такой review_settings не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)