+++ apps/api/views/review_views.py
@@ -10,8 +10,8 @@ from django.core.exceptions import ObjectDoesNotExist
 
 from ..serializers import ReviewSerializer
 
-from ..services.branch_service import get_branch_by_branch_id
-from ..services.review_service import create_review_by_branch_id, get_review_by_id
+from ..services.branch_service import get_branch_by_id
+from ..services.review_service import create_review_by_branch_id, get_review_by_id, get_all_review_by_company_id
 
 
 @swagger_auto_schema(
@@ -70,12 +70,7 @@ from ..services.review_service import create_review_by_branch_id, get_review_by_
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
 def create_review(request):
-    """
-    Контроллер для создания review
-    :param pk: id филиала
-    :param request:
-    :return: response
-    """
+    """Контроллер для создания review"""
     user = request.user
 
     try:
@@ -96,7 +91,7 @@ def create_review(request):
         return Response(data={'detail': 'Такой филиал не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -141,7 +136,7 @@ def read_review(request, pk):
         return Response(data={'detail': 'Такой review не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -175,7 +170,7 @@ def read_review_list(request, pk):
     user = request.user
 
     try:
-        branch = get_branch_by_branch_id(user=user, branch_id=pk)
+        branch = get_branch_by_id(user=user, branch_id=pk)
         if branch:
             serializer = ReviewSerializer(branch.review_set, many=True)
             return Response(data=serializer.data, status=status.HTTP_200_OK)
@@ -186,7 +181,7 @@ def read_review_list(request, pk):
         return Response(data={'detail': 'Такой филиал не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -265,7 +260,7 @@ def update_review(request, pk):
         return Response(data={'detail': 'Такой review не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
@@ -310,5 +305,50 @@ def delete_review(request, pk):
         return Response(data={'detail': 'Такой connect не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        message = 'Ошибка при обработке запроса ' + e.__str__()
+        message = f'Ошибка при обработке запроса {e}'
+        return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)
+
+
+@swagger_auto_schema(
+    method="get",
+    responses={
+        200: openapi.Response(
+            description='Запрос выполнен успешно',
+            schema=ReviewSerializer
+        ),
+        400: openapi.Response(
+            description='Ошибка при запросе'
+        ),
+        401: openapi.Response(
+            description='Пустой или неправильный токен'
+        ),
+        403: openapi.Response(
+            description='Ошибка доступа'
+        ),
+        404: openapi.Response(
+            description='Компания не найдена'
+        )
+    },
+    operation_description='Данный endpoint возвращает базовые данные о всех review по {id} компании.',
+    operation_summary='Получить информацию о reviews по id компании'
+)
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def read_review_list_all(request, pk):
+    """Контроллер для отдачи информации о всех review компании"""
+    user = request.user
+
+    try:
+        reviews = get_all_review_by_company_id(user=user, company_id=pk)
+        if reviews:
+            serializer = ReviewSerializer(reviews, many=True)
+            return Response(data=serializer.data, status=status.HTTP_200_OK)
+        else:
+            return Response(data={'detail': 'Это не ваша компания'}, status=status.HTTP_403_FORBIDDEN)
+
+    except ObjectDoesNotExist as er:
+        return Response(data={'detail': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
+
+    except Exception as e:
+        message = f'Ошибка при обработке запроса {e}'
         return Response(data={'detail': message}, status=status.HTTP_400_BAD_REQUEST)