+++ CHANGELOG.md
@@ -6,11 +6,15 @@
 
 ### Добавлено
 
--
+- Документация по всем слоям бизнес логики [@fringer2423](https://github.com/fringer2423)
+- Добавлен endpoint для возврата всех отзывов компании
+- Расширен слой бизнес логики для review
 
 ### Изменено
 
--
+- Fix Response detail performance [@fringer2423](https://github.com/fringer2423)
+- Fix Review serializer field (answered_at)
+- Refactor code
 
 ## [0.0.7] - 2022-10-13